Добавит модули util и data и перенесёт в них код из main
