+++ js/util.js
@@ -1,0 +1,28 @@
+const getRandomInteger = (a, b) => {
+  const min = Math.ceil(Math.min(a, b));
+  const max = Math.floor(Math.max(a, b));
+  return Math.floor(Math.random() * (max - min + 1) + min);
+};
+
+const getRandomArrayElement = (array) => array[getRandomInteger(0, array.length - 1)];
+
+const createRandomUniqueIdGenerator = (min, max) => {
+  const idValues = [];
+  return () => {
+
+    if (idValues.length >= (max - min + 1)) {
+      return null;
+    }
+
+    let newId = getRandomInteger(min, max);
+
+    while (idValues.includes(newId)) {
+      newId = getRandomInteger(min, max);
+    }
+
+    idValues.push(newId);
+    return newId;
+  };
+};
+
+export {getRandomInteger, getRandomArrayElement, createRandomUniqueIdGenerator};
